Move the python code into python files
